SG-5 Changed celestial object radius to double

+++ Core/Helpers/Shapes/ShapeHelper.vb
@@ -1,7 +1,7 @@
 ﻿Namespace Helpers.Shapes
     Public Class ShapeHelper
 
-        Public Shared Function VolumeOfASphere(radius As Integer) As Double
+        Public Shared Function VolumeOfASphere(radius As Double) As Double
             Return (4 / 3) * Math.PI * radius * radius
         End Function
